move duplicat key to internal package
