+++ libs/ibc-go/testing/simapp/encoding.go
@@ -2,7 +2,6 @@ package simapp
 
 import (
 	"github.com/okex/exchain/libs/cosmos-sdk/codec/types"
-	cosmoscryptocodec "github.com/okex/exchain/libs/cosmos-sdk/crypto/ibc-codec"
 	simappparams "github.com/okex/exchain/libs/ibc-go/testing/simapp/params"
 )
 
@@ -13,7 +12,7 @@ import (
 func MakeTestEncodingConfig() simappparams.EncodingConfig {
 	encodingConfig := simappparams.MakeTestEncodingConfig()
 	//std.RegisterLegacyAminoCodec(encodingConfig.Amino)
-	cosmoscryptocodec.RegisterInterfaces(encodingConfig.InterfaceRegistry)
+	codec.RegisterInterfaces(encodingConfig.InterfaceRegistry)
 	//ModuleBasics.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	interfaceReg := types.NewInterfaceRegistry()
 	ModuleBasics.RegisterInterfaces(interfaceReg)
